+++ app/src/main/java/com/bithack/apparatus/ModeMenu.java
@@ -78,8 +78,8 @@ public class ModeMenu extends Screen implements InputProcessor {
 
     @Override // com.badlogic.gdx.InputProcessor
     public boolean touchDown(int x, int y, int pointer, int btn) {
-        float px = x / G.realwidth;
-        float py = y / G.realheight;
+		float px = ((float) x) / ((float) G.realwidth);
+		float py = ((float) y) / ((float) G.realheight);
         Gdx.app.log("px,py", String.valueOf(px) + " " + py);
         if (x < 64 && y > G.realheight - 100) {
             this.tp.open_mainmenu();
